+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/open_trades/PeerProtocolLogMessageBox.java
@@ -20,7 +20,7 @@ package bisq.desktop.main.content.bisq_easy.open_trades;
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
 import bisq.desktop.main.content.components.chat_messages.ChatMessageListItem;
-import bisq.desktop.main.content.components.chat_messages.messages.MessageBox;
+import bisq.desktop.main.content.components.chat_messages.list_view.messages.MessageBox;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.control.Label;
