+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/components/chat_messages/list_view/messages/LeaveChatMessageBox.java
@@ -15,7 +15,7 @@
  * along with Bisq. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.main.content.components.chat_messages.messages;
+package bisq.desktop.main.content.components.chat_messages.list_view.messages;
 
 import bisq.chat.ChatChannel;
 import bisq.chat.ChatMessage;
